Added Boot to USB2UART
Added code to select between Access mode (the board enumerate as USB custom device and used as required by the user with the C# code) or USB2UART mode (board work as a simple USB <--> UART converter. 

To select access mode connect the board to a USB port and you will have 10 seconds to call FindDevHID method, this will keep the access mode, to use the USB2UART mode connect the board and wait more than 10 seconds, the board will reset and will be enumerated as a CDC device.